week12-yabu WebView内のリンクをタップすると外部ブラウザを開くようにした
