+++ app/src/main/kotlin/com/tumpaca/tumpaca/util/UIUtil.kt
@@ -1,9 +1,7 @@
 package com.tumpaca.tumpaca.util
 
-import android.util.Base64
 import android.webkit.WebView
-import android.webkit.WebViewClient
-import com.tumpaca.tumpaca.model.TPRuntime
+import com.tumpaca.tumpaca.view.TPWebViewClient
 
 /**
  * UIの便利メソッド集
@@ -12,42 +10,17 @@ import com.tumpaca.tumpaca.model.TPRuntime
 
 object UIUtil {
 
-    fun loadCss(webView: WebView, cssFile: String = "style.css"): Unit {
-        webView.settings.javaScriptEnabled = true
+    val TPWebViewClient = TPWebViewClient("style.css")
+    val DoNotHorizontalScrollWebViewClient = TPWebViewClient("doNotHorizontalScroll.css")
 
-        webView.setWebViewClient(object : WebViewClient() {
-            override fun onPageFinished(view: WebView, url: String) {
-                webView.loadUrl(generateCSS(cssFile))
-                super.onPageFinished(view, url)
-            }
-        })
+    fun loadCss(webView: WebView): Unit {
+        webView.settings.javaScriptEnabled = true
+        webView.setWebViewClient(TPWebViewClient)
     }
 
     fun doNotHorizontalScroll(webView: WebView): Unit {
-        loadCss(webView, "doNotHorizontalScroll.css")
+        webView.settings.javaScriptEnabled = true
+        webView.setWebViewClient(DoNotHorizontalScrollWebViewClient)
     }
-
-    private fun generateCSS(cssFile: String): String {
-        try {
-            val inputStream = TPRuntime.mainApplication.assets.open(cssFile)
-            val buffer = ByteArray(inputStream.available())
-            inputStream.read(buffer)
-            inputStream.close()
-            val encoded = Base64.encodeToString(buffer, Base64.NO_WRAP)
-            val scripts = "javascript:(function() {" +
-                    "var parent = document.getElementsByTagName('head').item(0);" +
-                    "var style = document.createElement('style');" +
-                    "style.type = 'text/css';" +
-                    // Tell the browser to BASE64-decode the string into your script !!!
-                    "style.innerHTML = window.atob('$encoded');" +
-                    "parent.appendChild(style)" +
-                    "})()"
-            return scripts
-        } catch (e: Exception) {
-            e.printStackTrace()
-            return ""
-        }
-    }
-
 
 }
